refactored common helper methods for working with cxf into a cxf-utils helper jar
