+++ components/cxf-utils/src/main/java/io/fabric8/utils/cxf/JsonHelper.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.repo.git;
+package io.fabric8.utils.cxf;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
