+++ components/fabric-utils/src/main/java/io/fabric8/utils/ssl/AllCertificatesTrustManager.java
@@ -13,7 +13,7 @@
  *  implied.  See the License for the specific language governing
  *  permissions and limitations under the License.
  */
-package io.fabric8.utils;
+package io.fabric8.utils.ssl;
 
 import java.security.cert.X509Certificate;
 import javax.net.ssl.X509TrustManager;
