+++ components/gitrepo-api/pom.xml
@@ -33,7 +33,7 @@
   <dependencies>
     <dependency>
       <groupId>io.fabric8</groupId>
-      <artifactId>fabric-utils</artifactId>
+      <artifactId>cxf-utils</artifactId>
     </dependency>
 
     <dependency>
