+++ components/gitrepo-api/src/test/java/io/fabric8/repo/git/Example.java
@@ -19,7 +19,7 @@ package io.fabric8.repo.git;
 
 import java.util.List;
 
-import static io.fabric8.repo.git.JsonHelper.toJson;
+import static io.fabric8.utils.cxf.JsonHelper.toJson;
 
 /**
  */
