+++ components/hubot-api/pom.xml
@@ -39,6 +39,10 @@
       <groupId>io.fabric8</groupId>
       <artifactId>fabric8-cdi</artifactId>
     </dependency>
+    <dependency>
+      <groupId>io.fabric8</groupId>
+      <artifactId>cxf-utils</artifactId>
+    </dependency>
 
     <dependency>
       <groupId>com.fasterxml.jackson.module</groupId>
